requirements.txt
PyQt6>=6.5
sounddevice>=0.4
numpy>=1.24
pynput>=1.7
google-cloud-speech>=2.22
